+++ libr/io/Makefile
@@ -8,7 +8,7 @@ R2DEPS+=r_cons
 R2DEPS+=r_crypto
 STATIC_OBJS=$(subst ..,p/..,$(subst io_,p/io_,$(STATIC_OBJ)))
 OBJS=${STATIC_OBJS}
-OBJS+=io.o io_plugin.o io_map.o io_desc.o io_cache.o p_cache.o undo.o ioutils.o io_fd.o
+OBJS+=io.o io_plugin.o io_map.o io_desc.o io_cache.o p_cache.o undo.o ioutils.o io_fd.o io_memory.o
 
 CFLAGS+=-Wall -DR2_PLUGIN_INCORE
